Conflicts:
	rosalina/source/menu.c
	rosalina/source/menus.c
	rosalina/source/menus/process_list.c

+++ rosalina/include/draw.h
@@ -25,7 +25,7 @@
 
 void draw_copyFramebuffer(void *dst);
 void draw_copyToFramebuffer(void *src);
-void draw_fillFramebuffer(u8 value);
+void draw_fillFramebuffer(u32 value);
 void draw_character(char character, u32 posX, u32 posY, u32 color);
 u32 draw_string(char *string, u32 posX, u32 posY, u32 color);
